Refactor file upload page into package
