+++ pages/file_upload/layout.py
@@ -1,0 +1,118 @@
+import dash_bootstrap_components as dbc
+from dash import dcc, html
+from components.advanced_upload import DragDropUploadArea
+from config.dynamic_config import dynamic_config
+
+def layout():
+    """File upload page layout with persistent storage"""
+    return dbc.Container(
+        [
+            # Removed redundant page header
+            # Upload area
+            dbc.Row(
+                [
+                    dbc.Col(
+                        [
+                            dbc.Card(
+                                [
+                                    dbc.CardHeader(
+                                        [html.H5("Upload Data Files", className="mb-0")]
+                                    ),
+                                    dbc.CardBody(
+                                        [
+                                            DragDropUploadArea(
+                                                id="upload-data",
+                                                max_size=dynamic_config.get_max_upload_size_bytes(),
+                                            ).render()
+
+                                        ]
+                                    ),
+                                ]
+                            )
+                        ]
+                    )
+                ]
+            ),
+            # Upload results area
+            dbc.Row([dbc.Col([html.Div(id="upload-results")])], className="mb-4"),
+            dbc.Row(
+                [
+                    dbc.Col(
+                        [
+                            dbc.Progress(
+                                id="upload-progress",
+                                value=0,
+                                label="0%",
+                                striped=True,
+                                animated=True,
+                            ),
+                            html.Ul(id="file-progress-list", className="list-unstyled mt-2")
+                        ]
+                    )
+                ],
+                className="mb-3",
+            ),
+            # Hidden button triggered by SSE when upload completes
+            dbc.Button("", id="progress-done-trigger", className="hidden"),
+            html.Div(id="sse-trigger", style={"display": "none"}),
+            # Data preview area
+            dbc.Row([dbc.Col([html.Div(id="file-preview")])]),
+            # Navigation to analytics
+            dbc.Row([dbc.Col([html.Div(id="upload-nav")])]),
+            # Container for toast notifications
+            html.Div(id="toast-container"),
+            # CRITICAL: Hidden placeholder buttons (using `.hidden` utility) to prevent callback errors
+            html.Div(
+                [
+                    dbc.Button("", id="verify-columns-btn-simple", className="hidden"),
+                    dbc.Button("", id="classify-devices-btn", className="hidden"),
+                ],
+                className="hidden",
+            ),
+            # Store for uploaded data info
+            dcc.Store(id="file-info-store", data={}),
+            dcc.Store(id="current-file-info-store"),
+            dcc.Store(id="current-session-id", data="session_123"),
+            dcc.Store(id="upload-task-id"),
+            dcc.Store(id="client-validation-store", data=[]),
+            dbc.Modal(
+                [
+                    dbc.ModalHeader(dbc.ModalTitle("Column Mapping")),
+                    dbc.ModalBody("Configure column mappings here", id="modal-body"),
+                    dbc.ModalFooter(
+                        [
+                            dbc.Button(
+                                "Cancel", id="column-verify-cancel", color="secondary"
+                            ),
+                            dbc.Button(
+                                "Confirm", id="column-verify-confirm", color="success"
+                            ),
+                        ]
+                    ),
+                ],
+                id="column-verification-modal",
+                is_open=False,
+                size="xl",
+            ),
+            dbc.Modal(
+                [
+                    dbc.ModalHeader(dbc.ModalTitle("Device Classification")),
+                    dbc.ModalBody("", id="device-modal-body"),
+                    dbc.ModalFooter(
+                        [
+                            dbc.Button(
+                                "Cancel", id="device-verify-cancel", color="secondary"
+                            ),
+                            dbc.Button(
+                                "Confirm", id="device-verify-confirm", color="success"
+                            ),
+                        ]
+                    ),
+                ],
+                id="device-verification-modal",
+                is_open=False,
+                size="xl",
+            ),
+        ],
+        fluid=True,
+    )
